safari-600.7-branch/LayoutTests/traversal/moz-mutations-3-expected.txt

CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B [E] * E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B [E] * E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root E [E1] * E11 (actual)
CONSOLE MESSAGE: line 69: root E [E1] * E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root E E1 [E11] * (actual)
CONSOLE MESSAGE: line 69: root E E1 [E11] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root E E1 [E11] * (actual)
CONSOLE MESSAGE: line 69: root E E1 [E11] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D E [E1] * E11 (actual)
CONSOLE MESSAGE: line 69: root B C D E [E1] * E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root C * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C n * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root C n * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root n2 C n * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root n2 C n * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C * [E] E1 E11 (actual)
CONSOLE MESSAGE: line 69: root C * [E] E1 E11 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C E E1 [E11] * (actual)
CONSOLE MESSAGE: line 69: root C E E1 [E11] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C [E] * (actual)
CONSOLE MESSAGE: line 69: root C [E] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C n [E] * (actual)
CONSOLE MESSAGE: line 69: root C n [E] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C n [E] * n2 (actual)
CONSOLE MESSAGE: line 69: root C n [E] * n2 (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C n E [n2] * (actual)
CONSOLE MESSAGE: line 69: root C n E [n2] * (expected)
CONSOLE MESSAGE: line 67: sequence check:
CONSOLE MESSAGE: line 68: root C n (actual)
CONSOLE MESSAGE: line 69: root C n (expected)
Results available in the JavaScript console
Test adapted from http://mxr.mozilla.org/mozilla-central/source/content/base/test/test_NodeIterator_mutations_3.html
